requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-telegram-bot==20.6
aiohttp==3.9.1
sqlalchemy==2.0.25
